feat: surport caculate totals in pivot data set
